Network diffusion

HierarchicalHotNet.stepmatrix — Function

Construct the step matrix for the random walk.

HierarchicalHotNet.random_walk_matrix — Function

Find the matrix that transforms given initial node probabilities into stationary distribution of visiting probabilities of a random walk with restart.

Missing docstring.

Missing docstring for HierarchicalHotNet.similarity_matrix. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HierarchicalHotNet.neighborhood_weights. Check Documenter's build log for details.

Network Permutation

One way to confirm the relevance of network diffusion-based predictions is to show that the results based on real the data differ significantly from the ones based on randomized data, where no meaningful patterns are expected. One scheme for generating realistic randomized dataset, proposed in the original paper, is to group the vertices with similar in- and out-degrees into bins and randomly shuffle the weights of the vertices within each bin. In this scheme, the weights of hubs are randomly reassigned to other hubs, and low-degree vertices to other low-degree vertices, so that overall the distribution of weights across the nodes looks very similar to real experiments, except that any correlations between the vertex weights and biological pathways are distrupted.

HierarchicalHotNet.vertexbins — Function

vertexbins(g::AbstractSimpleWeightedGraph,
           vertices::AbstractArray{Int}=1:nv(g);
           nbins::Integer=10, by::Symbol=:out, method=:tree) ->
    IndicesPartition

Partition the vertices of the weighted graph g into nbins bins, so that the vertices of the same bin have similar sum of edge weights.

Keyword arguments

nbins::Integer: the number of vertex bins. Use nbins &ge; 1 for bigger (nv &ge; 1) networks

by::Symbol: what edges to consider for grouping vertices. The supported options are:

:out (default, as in the original paper) – use only the outgoing edges

:in – use only the incoming edges

:outXin (recommended) – use all edges, so that the vertices of the same bin have similar sums of both incoming and outgoing edges

method::Symbol: the method for binning vertices

:sort (as in the original paper) – order the vertices by the sum of weights, then split the sorted array into equal bins. This method doesn't handle the by=:outXin well.

:tree (default) – build the hierarchical tree of vertices based on their sum of weights, then cut the tree to get the requested number of bins.

HierarchicalHotNet.randpermgroups! — Function

randpermgroups!(v::AbstractVector{Int}, groups::AbstractPartition) -> v

Randomly reshuffle the elements of groups within each group and put the result into v, group by group.

HierarchicalHotNet.randpermgroups — Function

randpermgroups(groups::AbstractPartition) -> Vector{Int}

Randomly reshuffle the elements of groups within each group.