Strongly Connected Components Tree

HierarchicalHotNet.SCCTree — Type

The tree of strongly connected components. Organizes the strongly connected components of the weighted directed graph into a tree.

Cutting the tree at specific threshold with cut gives the corresponding strongly connected components. The root of the tree corresponds to the weakest threshold.

HierarchicalHotNet.scctree — Function

scctree(g::Union{AbstractSimpleWeightedGraph, AbstractMatrix};
        method=:bisect, skipval=0, rev=false) -> SCCTree

Computes the hierarchical decomposition of the weighted directed graph g into strongly connected components.

Supports weighted graph objects as well as their adjacency matrix representations.

Keyword arguments

method::Symbol (defaults to :bisect): the method for partitioning. The supported methods are :bisect (the fastest) and :bottomup (slow, but simpler).

skipval::Number (defaults to zero): what value of the adjacency matrix should be treated as "no edge".

rev::Bool (defaults to false): if true, bigger edges weights are considered weaker than smaller ones

HierarchicalHotNet.cut — Function

cut(tree::SCCTree, threshold; minsize=1) -> IndicesPartition

Cuts the tree at the given edge weight threshold to get the corresponding strongly connected components of the original graph.

Keyword arguments

minsize: the minimal number of vertices in the component. Smaller connected components are skipped. By default returns components of any size.