Strongly Connected Components tree

HierarchicalHotNet.SCCTree — Type

The tree of strongly connected components. Organizes the strongly connected components of the weighted directed graph into a tree.

Cutting the tree at specific threshold with cut gives the corresponding strongly connected components. The root of the tree corresponds to the weakest threshold.

HierarchicalHotNet.scctree — Function

scctree(g::Union{AbstractSimpleWeightedGraph, AbstractMatrix};
        method=:bisect, skipval=0, rev=false) -> SCCTree

Computes the hierarchical decomposition of a weighted directed graph into strongly connected components.

Arguments

g::AbstractSimpleWeightedGraph: the graph (or its adjcency matrix representation) to partition into components

method::Symbol, defaults to :bisect: the method for partitioning. The supported methods are :bisect (the fastest) and :bottomup (slow, but simpler).

skipval::Number, defaults to zero: what value of the adjacency matrix should be treated as no edge.

rev::Bool, default to false: if true, bigger edges weights are considered weaker than smaller ones

HierarchicalHotNet.cut — Function

cut(tree, threshold; minsize=1) -> IndicesPartition

Cuts the tree at the given threshold to get the corresponding strongly connected components of the original graph. minsize optional parameter specifies whether components smaller than that would be ignored.