Source-Sink Analysis

For any edge weight threshold t, cutting the strongly connected components (SCC) tree provides the set of SCC of the original directed weighted network. Within each SCC subnetwork, there is a path from any node to any other node along the edges with weights not smaller than t. The original network may still contain the other edges with weights ≥t, which can connect one SCC subnetwork to the other, but it is not possible to reenter the same SCC again by traveling along these edges (otherwise by definition there is a bigger SCC). In other words, for each t we have a direct acyclic graph of connections between SCCs.

This property makes it very convenient to enumerate all paths from one set of nodes (sources) to the other (sinks). It is implemented by the HierarchicalHotNet.flowgraph() method. One can use it to identify signaling subnetworks that connect one biological data (e.g. interactors of a particular protein) to another (e.g. downstream changes resulting from knock out or overexpressing this protein).

The output of HierarchicalHotNet.flowgraph() is the subnetwork that consists of selected SCCs and the edges that connect these SCCs, plus the list of paths within this subnetwork from source to sink nodes. It could be shown that the path lengths tend to be smaller for the diffusion networks based on the real data than the ones based on the reshuffled node weights.

Missing docstring.

Missing docstring for HierarchicalHotNet.flowgraph. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HierarchicalHotNet.flowgraph!. Check Documenter's build log for details.

HierarchicalHotNet.nflows — Function

nflows(comps::IndicesPartition, adjmtx::AbstractMatrix,
       sources::AbstractVector{Int}, sinks::AbstractVector{Int},
       test::EdgeTest;
       maxweight::Union{Number, Nothing} = nothing,
       used_sources::Union{AbstractVector{Int}, Nothing}=nothing,
       used_sinks::Union{AbstractVector{Int}, Nothing}=nothing) -> NamedTupe

Calculates statistics from the flows from sources to sinks vertices in the weighted directed graph defined by adjmtx and test and the comps vertex components.

Returns the NamedTuple with the following fields

nflows: the number of source → sink flows

ncompoflows: the number of unique component(source) → component(sink) pairs for each source → sink flow

flowlen_sum: total length of all flows (flow length = the number of SCCs it crosses)

compflowlen_sum: total length of flows (every unique pairs of source/sink components is counted once)

flowinvlen_sum: the sum of $1/\mathrm{flowlength}$

compflowinvlen_sum: the sum of $1/\mathrm{flowlength}$ (each unique source/sink component pair is counted once)

compflowlen_max:

floweight_sum:

floweight_sum:

compfloweight_sum:

flowavghopweight_sum:

ncompsources: the number of distinct SCCs that have source nodes

ncompsinks: the number of distinct SCCs that have sink nodes

Keyword arguments

maxweight: if specified, limits the flow weight by maxweight

used_sources::AbstractVector{Int}: if specified, acts as an output parameter that contains the sorted list of sources that have outgoing flows

used_sinks::AbstractVector{Int}: if specified, acts as an output parameter that contains the sorted list of sinks that have incoming flows

HierarchicalHotNet.traceflows — Function

traceflows(step_adjmtx::AbstractMatrix, steptest::EdgeTest,
           walk_adjmtx::AbstractMatrix, walktest::EdgeTest;
           kwargs...) -> Dict{Diedge, Partition{Int}}

Trace the random walk (specified by walk_adjmtx and walktest) steps in the original graph (given by step_adjmtx and steptest).

See HierarchicalHotNet.traceflows! for the detailed description.

HierarchicalHotNet.traceflows! — Function

traceflows!(flow2paths::AbstractDict{Diedge, Partition{Int}},
            step_adjmtx::AbstractMatrix,
            steptest::EdgeTest,
            walk_adjmtx::AbstractMatrix,
            walktest::EdgeTest;
            sources::Union{AbstractVector, AbstractSet, Nothing} = nothing,
            sinks::Union{AbstractVector, AbstractSet, Nothing} = nothing,
            maxsteps::Integer=2) -> Dict{Diedge, Partition{Int}}

Trace the random walk (specified by walk_adjmtx and walktest) steps in the original graph (given by step_adjmtx and steptest).

Keyword arguments

sources (optional): the indices of vertices to use as path starts. If not specified, all vertices are used as path starts.

sinks (optional): the indices of vertices to use as path ends. If not specified, all vertices are used as path ends.

maxsteps=2: maximal number of steps in a traced path, longer paths are discarded.

Returns the mapping from the flow diedges to the HierarchicalHotNet.Parition object. Each part corresponds to the path, from diedge start to diedge end, in the original network, and the part elements are the indices of the intermedidate vertices along the path (start and end not included).