Network Statistics

Permutation allows generating randomized data that mimcs the key properties of the original vertex weight distribution. With enough permutations, it's possible to analyze how different are the results of network diffusion based on real weights in comparison to permuted weights.

The package allows doing this analysis at the level of individual vertices (HierarchicalHotNet.vertex_stats), directed edges (HierarchicalHotNet.diedge_stats), connected components (HierarchicalHotNet.conncomponents_stats) etc.

The statistcs from multiple permutation and cutting thresholds could be binned (HierarchicalHotNet.bin_treecut_stats) and then aggregated for calculating the quantiles of resulting distributions (HierarchicalHotNet.aggregate_treecut_binstats). Finally, HierarchicalHotNet.extreme_treecut_stats can find the edge cutting threshold with the maximal difference between the real and permuted weights.

HierarchicalHotNet.vertex_stats — Function

vertex_stats(weights::AbstractVector{<:Number},
             walkweights::AbstractVector{<:Number},
             permweights::AbstractMatrix{<:Number},
             walkpermweights::AbstractMatrix{<:Number}) -> DataFrame

Calculates statistics for the permuted vertex weights distribution and how it is different from the actual weights.

weights: weights of the vertices in the original network

walkweights: weights of the vertices after network diffusion analysis (stationary random walk distribution)

permweights: matrix of permuted weights; rows correspond to vertices, columns – to permutations

walkpermweights: matrix of vertex weights based on network diffusion analysis using permweights as input; rows correspond to vertices, columns to permutations

HierarchicalHotNet.diedge_stats — Function

diedge_stats(weights::AbstractVector{<:Number},
             walkweights::AbstractVector{<:Number},
             permweights::AbstractMatrix{<:Number},
             walkpermweights::AbstractMatrix{<:Number}) -> DataFrame

Calculates statistics for the directed edges permuted weights distribution and how it is different from the actual weights of directed edges.

Missing docstring.

Missing docstring for HierarchicalHotNet.conncomponents_stats. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HierarchicalHotNet.treecut_stats. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HierarchicalHotNet.treecut_compstats. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HierarchicalHotNet.bin_treecut_stats. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HierarchicalHotNet.aggregate_treecut_binstats. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HierarchicalHotNet.extreme_treecut_stats. Check Documenter's build log for details.