Partitions

Partition type provides an efficient container for storing multiple vectors of elements. The advantage over Vector{Vector{T}} is that internally it uses single Vector{T} to store the elements of all of its parts. The disadvantage is that it doesn't support adding or removing elements to/from arbitrary part, only the last part could be modified. Partition supports iterator interface for iterating over its parts as well as parts indexing and filter! for elements filtering.

HierarchicalHotNet.AbstractPartition — Type

Base class for partitions of elements of type T.

HierarchicalHotNet.Partition — Type

Efficiently stores the partition of a vector of elements of type T into disjoint sets.

Supports iterator interface.

HierarchicalHotNet.PartitionPart — Type

The type of parts returned by Partition{T}

HierarchicalHotNet.nelems — Function

nelems(ptn::AbstractPartition)

Total number of elements in the partition.

HierarchicalHotNet.elems — Function

elems(ptn::Partition{T}) where T -> Vector{T}

Vector of elements in the partition.

Rerturns the vectors as they are stored in ptn: the elements are ordered by their parts (elements from the first part come first etc) and maintain the same order as within each part.

HierarchicalHotNet.nparts — Function

nelems(ptn::AbstractPartition) -> Int

Number of parts in the partition.

Same as length(ptn).

Missing docstring.

Missing docstring for HierarchicalHotNet.partrange. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HierarchicalHotNet.partlength. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HierarchicalHotNet.ispartempty. Check Documenter's build log for details.

HierarchicalHotNet.pushelem! — Function

pushelem!(ptn::Partition, el) -> ptn

Push the element into the last part of partition.

HierarchicalHotNet.closepart! — Function

closepart!(ptn::Partition) -> ptn

Finalize the last part of the partition and append the new empty part.

All subsequent calls to pushelem! will add elements into the new part.

HierarchicalHotNet.repeat! — Function

repeat!(ptn::Partition, n::Integer) -> ptn

Repeat the parts of ptn n times.

julia> using HierarchicalHotNet

julia> ptn = HierarchicalHotNet.IndicesPartition(5, nparts=1)
1-element HierarchicalHotNet.Partition{Int64}:
 [1, 2, 3, 4, 5]

julia> HierarchicalHotNet.repeat!(ptn, 3)
3-element HierarchicalHotNet.Partition{Int64}:
 [1, 2, 3, 4, 5]
 [1, 2, 3, 4, 5]
 [1, 2, 3, 4, 5]

HierarchicalHotNet.IndicesPartition — Type

Partition of an integer vector

HierarchicalHotNet.reset! — Function

reset!(p::IndicesPartition, n::Integer=nelems(p); nparts::Integer=n) -> p

Resets the integer partition p.

If nparts == 1, the partition is reset into [[1, 2, 3, ..., n]]. If n == nparts, sets the partition to [[1], [2], [3], ..., [n]].